requirements.txt
allure-pytest==2.13.5
dotenv==0.9.9
iniconfig==2.1.0
playwright==1.51.0
pytest==8.3.5
pytest-asyncio==1.0.0
pytest-base-url==2.1.0
pytest-playwright==0.7.0
python-dotenv==1.1.0
requests==2.32.3
selenium==4.30.0
web3==7.12.0
websocket-client==1.8.0
websockets==15.0.1

